+++ docs/demo/dragdrop/drop.tsx
@@ -5,6 +5,11 @@ import FlowBuilder, {
   IRegisterNode,
   IDropComponent,
 } from 'react-flow-builder';
+import {
+  DrawerComponent,
+  PopconfirmComponent,
+  PopoverComponent,
+} from '../antd';
 
 import './index.css';
 
@@ -133,6 +138,9 @@ const Dragdrop = () => {
       nodes={nodes}
       onChange={handleChange}
       registerNodes={registerNodes}
+      DrawerComponent={DrawerComponent}
+      PopoverComponent={PopoverComponent}
+      PopconfirmComponent={PopconfirmComponent}
     />
   );
 };